Added comment explaining why a velocity PIDSource returns kDisplacement
In order for the P constant of a PID controller to produce an output
proportional to the setpoint error, the PIDSource type must be
kDisplacement (otherwise, the P term is effectively remapped to an I term).
To work around this, the VelocitySrouce toDisplacementPIDSource creates
a PIDSource that always returns kDisplacement as the type.